% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EYR.r
\docType{data}
\name{EYR}
\alias{EYR}
\title{Eastern yellow robin (EYR) DArT data in a genlight object}
\format{
genlight object
}
\usage{
EYR
}
\description{
Genetic data from EYR collected in VIC, Australia by researchers at Monash 
University. The data comes in the form of a genlight object ready to be used 
by dartR. The file contains data for 782 individuals from 4 populations and
1,000 binary SNPs. The dataset contains a mix of autosomal and sex-linked
loci.
}
\references{
Robledo-Ruiz et al. (2023) Easy-to-use R functions to separate 
reduced-representation genomic datasets into sex-linked and autosomal loci, 
and conduct sex assignment. \emph{Mol. Ecol. Res.}
(\url{https://doi.org/10.1111/1755-0998.13844})
}
\author{
Custodian: Diana Robledo-Ruiz -- Post to
  \url{https://groups.google.com/d/forum/dartr}
}
\keyword{datasets}
