% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.drop.sexlinked.r
\name{gl.drop.sexlinked}
\alias{gl.drop.sexlinked}
\title{Used to remove loci that are sex linked}
\usage{
gl.drop.sexlinked(x, system = NULL)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP data. This genlight
object needs to contain the sex of the individuals [required].}

\item{system}{String that declares the sex-determination system of the 
species: 'zw' or 'xy' [required].}
}
\value{
A warning.
}
\description{
This function is deprecated. Please use \code{\link{gl.filter.sexlinked}} 
instead.
}
\details{
This function has been deprecated and replaced by gl.filter.sexlinked in 
order to keep consistency with other functions (gl.report -> gl.filter).
}
\references{
\itemize{
\item Robledo‐Ruiz, D. A., Austin, L., Amos, J. N., Castrejon‐Figueroa, J.,
 Harley, D. K., Magrath, M. J., Sunnucks, P., & Pavlova, A. (2023). 
 Easy‐to‐use R functions to separate reduced‐representation genomic datasets
  into sex‐linked and autosomal loci, and conduct sex assignment. Molecular 
  Ecology Resources, 00, 1-21.
 }
}
\author{
Custodian: Diana Robledo-Ruiz -- Post to
  \url{https://groups.google.com/d/forum/dartr}
}
