% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.filter.sexlinked.r
\name{gl.filter.sexlinked}
\alias{gl.filter.sexlinked}
\title{Removes loci that are sex linked}
\usage{
gl.filter.sexlinked(
  x,
  system = NULL,
  ncores = 1,
  plot.display = TRUE,
  plot.theme = theme_dartR(),
  plot.colors = NULL,
  plot.file = NULL,
  plot.dir = NULL,
  verbose = NULL
)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP data. This genlight
object needs to contain the sex of the individuals. See explanation in
details [required].}

\item{system}{String that declares the sex-determination system of the
species: 'zw' or 'xy' [required].}

\item{ncores}{Number of processes to be used in parallel operation. If ncores
> 1 parallel operation is activated [default 1].}

\item{plot.display}{If TRUE, displays the four output plots. See
explanation in details [default TRUE].}

\item{plot.theme}{Theme for the plots [default theme_dartR()].}

\item{plot.colors}{Not implemented; the plots use fixed colours for each
category of loci [default NULL].}

\item{plot.file}{Name for the RDS binary file to save (base name only,
exclude extension) [default NULL].}

\item{plot.dir}{Directory to save the plot RDS files [default as specified
by the global working directory or tempdir()].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
progress log; 3, progress and results summary; 5, full report
[default 2, unless specified using gl.set.verbosity].}
}
\value{
A genlight object with autosomal loci, or NULL (see Details).
}
\description{
This function identifies sex-linked and autosomal loci present in a SNP
dataset (genlight object) using individuals with known sex. It identifies
five types of loci: w-linked or y-linked, sex-biased, z-linked or
x-linked, gametologous and autosomal.

This function returns a genlight object with autosomal loci only, and
displays 4 plots.
}
\details{
The genlight object must contain in \code{gl@other$ind.metrics} a column
named "sex" (case-insensitive) in which individuals with known sex are
assigned 'F' or 'Female' for females, and 'M' or 'Male' for males
(case-insensitive). The rows of \code{ind.metrics} must be in the same
order as the individuals in the genlight object. The function ignores
individuals that are assigned anything else or nothing at all
(unknown sex). At least one female and one male are required. SilicoDArT
(presence/absence) data are not accepted.

Loci are classified as in \code{\link{gl.report.sexlinked}}, which
describes the tests and thresholds used.

The display of plots can be turned off (\code{plot.display = FALSE}), but
we strongly encourage you to always inspect the output plots at least once
to make sure everything is working properly.

\strong{ Function's output }

This function returns a genlight object that contains only autosomal loci
(i.e. sex-linked loci have been filtered out), or NULL if every locus is
sex-linked. The object keeps all individuals, carries the locus metrics of
the autosomal loci only, and records this call in its history. To keep the
sex-linked loci instead, use \code{\link{gl.keep.sexlinked}}.

And displays four plots:\itemize{
\item {A BEFORE plot based on loci call rate by sex, with w/y-linked loci
 colored in yellow and sex-biased loci in blue}
\item {An AFTER plot based on loci call rate by sex, with sex-linked loci
removed}
\item {A BEFORE plot based on loci heterozygosity by sex, with z/x-linked
loci colored in orange and gametologs in green}
\item {An AFTER plot based on loci heterozygosity by sex, with sex-linked
loci removed}
}
The plots are not returned; use \code{plot.file} to save them.
}
\examples{
LBP_noSexLinked <- gl.filter.sexlinked(x = LBP, system = "xy", 
plot.display = TRUE, ncores = 1)
LBP_noSexLinked

}
\references{
\itemize{
\item Robledo-Ruiz, D. A., Austin, L., Amos, J. N., Castrejon-Figueroa, J.,
 Harley, D. K., Magrath, M. J., Sunnucks, P., & Pavlova, A. (2023).
 Easy to use R functions to separate reduced representation genomic datasets
  into sex linked and autosomal loci, and conduct sex assignment. Molecular
  Ecology Resources, 00, 1-21.
 }
}
\author{
Author(s): Diana Robledo-Ruiz. Custodian: Diana Robledo-Ruiz --
  Post to \url{https://groups.google.com/d/forum/dartr}
}
\concept{matched filter}
