% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.sexlinked.r
\name{gl.report.sexlinked}
\alias{gl.report.sexlinked}
\title{Reports loci that are sex linked}
\usage{
gl.report.sexlinked(
  x,
  system = NULL,
  ncores = 1,
  plot.display = TRUE,
  plot.theme = theme_dartR(),
  plot.colors = NULL,
  plot.file = NULL,
  plot.dir = NULL,
  verbose = NULL
)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP data. This genlight
object needs to contain the sex of the individuals. See explanation in
details [required].}

\item{system}{String that declares the sex-determination system of the
species: 'zw' or 'xy' [required].}

\item{ncores}{Number of processes to be used in parallel operation. If ncores
> 1 parallel operation is activated [default 1].}

\item{plot.display}{If TRUE, displays the two output plots. See explanation
in details [default TRUE].}

\item{plot.theme}{Theme for the plots [default theme_dartR()].}

\item{plot.colors}{Not implemented; the plots use fixed colours for each
category of loci [default NULL].}

\item{plot.file}{Name for the RDS binary file to save (base name only,
exclude extension) [default NULL].}

\item{plot.dir}{Directory to save the plot RDS files [default as specified
by the global working directory or tempdir()].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
progress log; 3, progress and results summary; 5, full report
[default 2, unless specified using gl.set.verbosity].}
}
\value{
A dataframe with one row per locus (see Details).
}
\description{
This function identifies sex-linked and autosomal loci present in a SNP
dataset (genlight object) using individuals with known sex. It identifies
five types of loci: w-linked or y-linked, sex-biased, z-linked or
x-linked, gametologous and autosomal.

This function returns a dataframe and displays 2 plots.
}
\details{
The genlight object must contain in \code{gl@other$ind.metrics} a column
named "sex" (case-insensitive) in which individuals with known sex are
assigned 'F' or 'Female' for females, and 'M' or 'Male' for males
(case-insensitive). The rows of \code{ind.metrics} must be in the same
order as the individuals in the genlight object. The function ignores
individuals that are assigned anything else or nothing at all
(unknown sex). At least one female and one male are required. SilicoDArT
(presence/absence) data are not accepted.

\strong{ Classification of loci }

For each locus, the function tests for independence of sex and call rate
(phase 1), and of sex and heterozygosity (phase 2). It uses Fisher's exact
test when fewer than 1,000 individuals of known sex are scored, and a
chi-square test (without continuity correction) otherwise. P-values are
adjusted for false discovery rate. Loci are then classified as follows:
\itemize{
\item {w-linked (zw) or y-linked (xy): adjusted p-value of phase 1 <= 0.01
and call rate <= 0.1 in males (zw) or females (xy).}
\item {sex-biased: adjusted p-value of phase 1 <= 0.01, and not w/y-linked.}
\item {z-linked (zw) or x-linked (xy): not w/y-linked or sex-biased,
adjusted p-value of phase 2 <= 0.01, and heterozygosity higher in males
(zw) or females (xy).}
\item {gametolog: as z/x-linked, but heterozygosity not higher in males
(zw) or females (xy).}
\item {autosomal: all other loci.}
}

The display of plots can be turned off (\code{plot.display = FALSE}), but
we strongly encourage you to always inspect the output plots at least once
to make sure everything is working properly.

\strong{ Function's output }

This function displays two plots:\itemize{
\item {A plot based on loci call rate by sex, with w/y-linked loci colored
       in yellow and sex-biased loci in blue}
\item {A plot based on loci heterozygosity by sex, with z/x-linked loci
 colored in orange and gametologs in green}
}
The plots are not returned; use \code{plot.file} to save them.

And it returns a dataframe in which loci are in rows, and columns:\itemize{
\item {index - Index number to identify loci}
\item {count.F.miss - Count of females that have this locus as missing
data (NA).}
\item {count.M.miss - Count of males that have this locus as missing
data (NA)}
\item {count.F.scored - Count of females that have this locus scored
(0, 1 or 2; i.e. non-missing)}
\item {count.M.scored - Count of males that have this locus scored
(0, 1 or 2; i.e. non-missing)}
\item {ratio - Test statistic for the independence of call rate and sex
for this locus: the Fisher's exact test odds ratio estimate, or the
chi-square statistic when 1,000 or more individuals are tested}
\item {p.value - P-value of the test}
\item {p.adjusted - P-value adjusted for false discovery rate}
\item {scoringRate.F - Female call rate (proportion of females that were
scored for this locus; x-axis in the 1st plot)}
\item {scoringRate.M - Male call rate (proportion of males that were scored
for this locus; y-axis in the 1st plot)}
\item {w.linked/y.linked - Boolean for this locus being w-linked/y-linked}
\item {sex.biased - Boolean for this locus having sex-biased call rate}
\item {count.F.het - Count of females that are heterozygous for this locus}
\item {count.M.het - Count of males that are heterozygous for this locus}
\item {count.F.hom - Count of females that are homozygous for this locus}
\item {count.M.hom - Count of males that are homozygous for this locus}
\item {stat - Test statistic for the independence of heterozygosity and
sex for this locus: the Fisher's exact test odds ratio estimate, or the
chi-square statistic when 1,000 or more individuals are tested. NA for
w/y-linked and sex-biased loci, which are not tested}
\item {stat.p.value - P-value of the test}
\item {stat.p.adjusted - P-value adjusted for false discovery rate}
\item {heterozygosity.F - Proportion of females that are heterozygotes for
this locus (x-axis in the 2nd plot)}
\item {heterozygosity.M - Proportion of males that are heterozygotes for
this locus (y-axis in the 2nd plot)}
\item {z.linked/x.linked - Boolean for this locus being z-linked/x.linked}
\item {gametolog - Boolean for this locus being a gametolog}
}
}
\examples{
out <- gl.report.sexlinked(x = LBP, system = "xy", plot.display = TRUE,
ncores = 1)

}
\references{
\itemize{
\item Robledo-Ruiz, D. A., Austin, L., Amos, J. N., Castrejon-Figueroa, J.,
 Harley, D. K., Magrath, M. J., Sunnucks, P., & Pavlova, A. (2023).
 Easy to use R functions to separate reduced representation genomic datasets
  into sex linked and autosomal loci, and conduct sex assignment. Molecular
  Ecology Resources, 00, 1-21.
 }
}
\author{
Author(s): Diana Robledo-Ruiz. Custodian: Diana Robledo-Ruiz --
  Post to \url{https://groups.google.com/d/forum/dartr}
}
\concept{matched report}
