% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.filter.sexlinked.r
\name{gl.filter.sexlinked}
\alias{gl.filter.sexlinked}
\title{Filters loci that are sex linked}
\usage{
gl.filter.sexlinked(gl, system = NULL, plots = TRUE, ncores = 1)
}
\arguments{
\item{gl}{Name of the genlight object containing the SNP data. This genlight
object needs to contain the sex of the individuals. See explanation in 
details [required].}

\item{system}{String that declares the sex-determination system of the 
species: 'zw' or 'xy' [required].}

\item{plots}{Creates four output plots. See explanation in details
[default TRUE].}

\item{ncores}{Number of processes to be used in parallel operation. If ncores
> 1 parallel operation is activated, see "Details" section [default 1].}
}
\value{
A list of 6 elements and 4 plots.
}
\description{
This function identifies sex-linked and autosomal loci present in a SNP 
dataset (genlight object) using individuals with known sex. It identifies
five types of loci: w-linked or y-linked, sex-biased, z-linked or
x-linked, gametologous and autosomal.

This function produces as output a list with 6 elements and 4 plots.
}
\details{
The genlight object must contain in \code{gl@other$ind.metrics} a column 
named "id", and a column named "sex" in which individuals with known-sex are 
assigned 'M' for male, or 'F' for female. The function ignores individuals 
that are assigned anything else or nothing at all (unknown-sex).

The creation of plots can be turned-off (plots = FALSE) in order to save a 
little bit of running time for very large datasets (>50,000 SNPs). However, 
we strongly encourage you to always inspect the output plots at least once to
make sure everything is working properly.

\strong{ Function's output }

This function creates a list of 6 elements: \itemize{
\item {$results.table > Table with statistics (columns) for each loci (rows)}
\item {$w.linked or $y.linked > Genlight object with w-linked/y-linked loci}
\item {$sex.biased > Genlight object with sex-biased scoring rate loci} 
\item {$z.linked or $x.linked > Genlight object with z-linked/x-linked loci}
\item {$gametolog > Genlight object with gametologs} 
\item {$autosomal > Genlight object with autosomal loci}
}

And 4 plots:\itemize{
\item {A plot BEFORE filtering sex-linked loci by call rate}
\item {A plot AFTER filtering sex-linked loci by call rate} 
\item {A plot BEFORE filtering sex-linked loci by heterozygosity}
\item {A plot AFTER filtering sex-linked loci by heterozygosity}
}
}
\examples{
out <- gl.filter.sexlinked(gl = LBP, system = "xy", plots = TRUE, ncores = 2)
out$results.table
out$x.linked

}
\author{
Custodian: Diana Robledo-Ruiz -- Post to
  \url{https://groups.google.com/d/forum/dartr}
}
